Minor: move `AnalysisContext` out of physical_expr and into its own module (#7127)
